requirements.txt
python-telegram-bot
configparser
oauth2
cryptography
pymongo
